Fix one sign-compare error in OCTET_STRING.c
OCTET_STRING.c:1896:46: error: signed and unsigned type in conditional expression [-Werror=sign-compare]

+++ skeletons/OCTET_STRING.c
@@ -1893,7 +1893,7 @@ OCTET_STRING_random_length_constrained(
         if(pc->flags & APC_CONSTRAINED) {
             long suggested_upper_bound = pc->upper_bound < (ssize_t)max_length
                                              ? pc->upper_bound
-                                             : max_length;
+                                             : (ssize_t)max_length;
             if(max_length <= (size_t)pc->lower_bound) {
                 return pc->lower_bound;
             }
